Micro-optimize the combineDF part of NameConflictTreeWalk
We can do the common case of combineDF (which requires no lookahead)
during the search for the minimum entry by keeping track of the type
of each entry and using the single entry lookahead inherent in the
data structure.

Since this catches the most common case of "a" and "a/" with no
intervening "a.foo" we can typically avoid the much more costly
combineDF and skipEntry routines.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>