Checks added to avoid a lot of entries in the log
IResource.isAccessible() must be checked before calling setProperty on it.
Happend When opening a java project shared with CVS (or synchronizing the same project),
egit throws exception for files he tries to decorate wrongly.

GitResourceDecorator.getActiveDecorator() was returning null when opening a
java project shared with CVS (or synchronizing the same project)

[Edited description and use isAccessible instead of exists to test for applicability.
The latter is a partial revert of 4a230ea1]

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>