JUnit test for FileTreeIterator, WorkingTreeIterator
This test is an indirect test of WorkingTreeIterator as the class
is abstract.  An easy way to test it is through the FileTreeIterator
as that relies only on standard Java IO.

We don't have full coverage yet, but most of the cases are dealt
with by this test.  A better test of WorkingTreeIterator would be
to setup a mock implementation that works strictly in memory and
throws it all of the corner cases.  Building that is a lot more
work, and the test will be focusing more on the mock code than on
the live library, so I'm punting on it for now.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>