Build jgit.jar and jgit_src.zip alongside jgit CLI wrapper
When we build jgit the CLI executable we have the classes already
compiled so we can easily construct a jgit.jar and jgit_src.zip
at the same time.  This makes it easier for folks who need the
library and not the command line interface package.

We do not include JSch in jgit.jar as we assume the downstream
user can supply us the package.  In an IDE scenario it is quite
likely the IDE already has a copy of JSch available for use, so
we really don't want to supply our own and potentially put two
different versions on the classpath.

The version we embed within our JAR manifest is now more like
the dotted format used by C Git or git-gui.  In particular we
embed the "-dirty" suffix if the tree contains modifications.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>