Introduce AnyObjectId.name() to replace standard toString()
This allows objects that extend AnyObjectId to override toString() to
their needs. With this it is possible to subclass for example RevCommit
and provide a custom toString() implementation without breaking the
revision walker.

Only the new method is introduced.  Existing callers still must be
modified to use name() instead of an explicit or implicit toString().

Signed-off-by: Jonas Fonseca <fonseca@diku.dk>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>