Cleanup RevWalk.parseTree semantics
When we call parseAny we may have opened the ObjectLoader for the
tree we are going after.  If that worked we know the object exists
in the repository and is certainly recorded as type "tree" so we
do not need to open the object a second time before returning to
the caller.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>