Add prerequisite verification to the bundle transport
Before we can fetch the objects contained within the bundle into
the repository we need to know if we have the necessary bases in
our reachable history, thus ensuring that we have what is needed
to fill out the thin pack and access everything contained within
it.  If we are missing any of the commits named as prerequisites
in the bundle header we cannot import it.

I forgot to implement this when I initially added bundle support
to the transport implementation.  Better late then never.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>