Relocate test resources to classpath and load from there
Previously test case resources were located in the root directory, this
change relocates them to a classpath location and loads the resources from
there.

If there is request for any test resource that is not present in classpath
then util method will return the same value as it would in previous state.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>