Paper bag fix RevWalk.reset after inMergeBase is used
We need to remove delayedFreeFlags from carryFlags anytime we
mark those delayedFreeFlags as actually freeFlags.  In other
words we do not want to continue carrying a flag which we have
now freed and will recycle for a different use in the future,
one which may not want to be carried automatically onto parent
commits during revision traversal.

I had the boolean expression incorrect (call it a typo).  The
correct way to remove set b from a is "a &= ~b" not "a &= b".

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>