jgit programs: Use i18n.logOutputEncoding or user's locale for output
We should present the data in the user's locale to make it readable.

This prevents garbage from being displayed for user's whose locale
is not UTF-8, when non-ascii appears in most cases. If the characters
cannot be converted garbage will appear in any case.

When jgit gains the capability to present blob data this code must
be replaced to handle data and metadata differently.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>