Patch parse test comparing "git log -p" output to "git log --numstat"
By comparing the output of "git log -p", once parsed by our patch
parser class, to the output of "git log --numstat" we can be quite
certain we are reading the patches from Git with a high degree of
accuracy, at least for typical add/remove sorts of changes (no
rename detection).

Unfortunately two commits in our history produce an off-by-one bug
in git log --numstat.  The bug appears to be in log --numstat and
not in JGit as git apply --numstat matches JGit's result, and is
thus also differing from log --numstat.  Since this occurs on only
2 commits out of 1,211 processed during the test I'm not worrying
about the difference on these two items.  Besides the numbers from
JGit and git apply --numstat look to be more correct.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>