Refactor the old/pre-image data in HunkHeader to support >1 ancestor
The "diff --cc" format uses more than one ancestor in each hunk,
so we need to expand the hunk header information in a way that
allows access to the data for each ancestor.  This change moves
the information relative to the old/pre-image ancestor in a 2-way
patch into an OldImage object.  In a 2-way patch we only have
one OldImage, but in a "diff --cc" patch we will have more than
one of these available in each hunk.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>