Use FileMode.MISSING when a file is added or deleted rather than null
Null is better used to indicate "no mode information at all" in the
patch, while FileMode.MISSING is already commonly used within the
TreeWalk code to mean "this path doesn't exist in this tree".  In
the context of a patch to create or delete a file, MISSING makes a
lot more sense for the application to work with.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>