Mark locations that break in the year 2038
When the 32 bit timestamp rolls over in 2038 these spots in JGit
and everything that depends on them will break, unless we upgrade
them to use a 64 bit long before then. For now its too time and memory
intensive to use a full long here,but in another 10 years we should
have enough computing power that this is a moot point and we can upgrade
the code paths marked with by 'git grep 2038'.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>