Allow null new ObjectId during RefUpdate.delete
If we are deleting a ref we really don't care about what the
new ObjectId value should be; it just doesn't matter.  If we
didn't set the value we should consider it the same as if we
had a MissingObjectException, which means treat the value as
null and perform a force update test.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>