Teach TransportLocal how to use our own ReceivePack implementation
If the user hasn't reconfigured git receive-pack on this remote
then we can use our internal implementation instead of forking
out to the C implementation.  It may be faster to stay within
the JVM then to start the external process.

In-memory pipes are constructed to copy the data between the
calling client thread and the background ReceivePack worker.
In certain cases we may even be able to do a faster copy by
directly hard-linking object files and/or packs, but that
isn't traditionally how Git behaves, so we use the standard
network protocol anyway.

Since our ReceivePack process doesn't fire hooks by default
we no longer will invoke any hooks during a push to a local
repository.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>